+++ src/Uno.UI.Composition/Composition/CompositionViewBox.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 
 using System.Numerics;
+using Windows.Foundation;
 
 namespace Microsoft.UI.Composition;
 
@@ -45,4 +46,11 @@ public partial class CompositionViewBox : CompositionObject
 		get => _horizontalAlignmentRatio;
 		set => SetProperty(ref _horizontalAlignmentRatio, value);
 	}
+
+	internal Rect GetRect()
+		=> new(
+			x: Offset.X,
+			y: Offset.Y,
+			width: Size.X,
+			height: Size.Y);
 }
